+++ src/test/java/com/jagrosh/jmusicbot/unit/BotConfigUnitTest.java
@@ -417,6 +417,7 @@ class BotConfigUnitTest extends BaseConfigTest {
                 discord.owner = 123456789
                 playback.audioSources {
                   youtube = true
+                  spotify = false
                   soundcloud = true
                   bandcamp = false
                   vimeo = true
@@ -436,7 +437,7 @@ class BotConfigUnitTest extends BaseConfigTest {
             
             Set<AudioSource> sources = config.getEnabledAudioSources();
             
-            // Should have 8 sources enabled (10 total - 2 disabled)
+            // Should have 8 sources enabled (11 total - 3 disabled: bandcamp, beam, spotify)
             assertEquals(8, sources.size(), "Should have 8 sources enabled");
             
             // Enabled sources
@@ -454,6 +455,8 @@ class BotConfigUnitTest extends BaseConfigTest {
                 "Bandcamp should be disabled when set to false");
             assertFalse(config.isAudioSourceEnabled(AudioSource.BEAM),
                 "Beam should be disabled when set to false");
+            assertFalse(config.isAudioSourceEnabled(AudioSource.SPOTIFY),
+                "Spotify should be disabled when set to false");
         }
         
         @Test
@@ -467,6 +470,7 @@ class BotConfigUnitTest extends BaseConfigTest {
                 discord.owner = 123456789
                 playback.audioSources {
                   youtube = false
+                  spotify = false
                   soundcloud = false
                   bandcamp = false
                   vimeo = false
@@ -652,6 +656,7 @@ class BotConfigUnitTest extends BaseConfigTest {
                 discord.owner = 123456789
                 playback.audioSources {
                   youtube = false
+                  spotify = false
                   soundcloud = true
                   bandcamp = false
                   vimeo = true